--- pekko/src/main/java/com/tok/pekko/common/actor/CborSerializable.java
@@ -1,4 +1,0 @@
-package com.tok.pekko.common.actor;
-
-public interface CborSerializable {
-}
