+++ pekko/src/main/java/com/tok/pekko/infrastructure/actor/CborSerializable.java
@@ -1,0 +1,4 @@
+package com.tok.pekko.infrastructure.actor;
+
+public interface CborSerializable {
+}
